pyproject.toml
[project]
name = "mail_collector"
version = "0.0.1"
description = "mail_collector for hackatone"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi>=0.122.0",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pre-commit>=4.5.0",
    "ruff>=0.14.6",
]

[tool.ruff]
line-length = 79
target-version = 'py312'
exclude = ['tests/']

[tool.ruff.lint]
select = ["E", "F", "I", "B", "Q"]
fixable = ['ALL']
ignore = ['E101']

[tool.ruff.lint.flake8-quotes]
inline-quotes = 'single'
multiline-quotes = "double"

[tool.ruff.format]
quote-style = 'single'

[tool.mypy]
python_version = '3.12'
disallow_untyped_defs = true
strict_optional = true
exclude = [
    'tests/',
    '.venv/',
    '__pycache__'
]
ignore_missing_imports = true


[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.coverage.run]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__init__.py",
    "*/src/config/consts.py"
]
